PMM-12378 Extract services table
