+++ public/app/percona/inventory/Tabs/Services.tsx
@@ -5,15 +5,10 @@ import { Row } from 'react-table';
 
 import { AppEvents } from '@grafana/data';
 import { locationService } from '@grafana/runtime';
-import { Badge, Button, HorizontalGroup, Icon, Link, Modal, TagList, useStyles2 } from '@grafana/ui';
+import { Button, HorizontalGroup, Modal, useStyles2 } from '@grafana/ui';
 import { OldPage } from 'app/core/components/Page/Page';
-import { stripServiceId } from 'app/percona/check/components/FailedChecksTab/FailedChecksTab.utils';
-import { Action } from 'app/percona/dbaas/components/MultipleActions';
 import { CheckboxField } from 'app/percona/shared/components/Elements/Checkbox';
-import { DetailsRow } from 'app/percona/shared/components/Elements/DetailsRow/DetailsRow';
 import { FeatureLoader } from 'app/percona/shared/components/Elements/FeatureLoader';
-import { ServiceIconWithText } from 'app/percona/shared/components/Elements/ServiceIconWithText/ServiceIconWithText';
-import { ExtendedColumn, FilterFieldTypes, Table } from 'app/percona/shared/components/Elements/Table';
 import { FormElement } from 'app/percona/shared/components/Form';
 import { useCancelToken } from 'app/percona/shared/components/hooks/cancelToken.hook';
 import { usePerconaNavModel } from 'app/percona/shared/components/hooks/perconaNavModel';
@@ -24,28 +19,17 @@ import {
 } from 'app/percona/shared/core/reducers/services';
 import { getServices } from 'app/percona/shared/core/selectors';
 import { isApiCancelError } from 'app/percona/shared/helpers/api';
-import { getDashboardLinkForService } from 'app/percona/shared/helpers/getDashboardLinkForService';
-import { getExpandAndActionsCol } from 'app/percona/shared/helpers/getExpandAndActionsCol';
 import { logger } from 'app/percona/shared/helpers/logger';
-import { ServiceStatus } from 'app/percona/shared/services/services/Services.types';
 import { useAppDispatch } from 'app/store/store';
 import { useSelector } from 'app/types';
 
 import { appEvents } from '../../../core/app_events';
 import { GET_SERVICES_CANCEL_TOKEN } from '../Inventory.constants';
 import { Messages } from '../Inventory.messages';
-import { FlattenService, MonitoringStatus } from '../Inventory.types';
-import { StatusBadge } from '../components/StatusBadge/StatusBadge';
-import { StatusInfo } from '../components/StatusInfo/StatusInfo';
-import { StatusLink } from '../components/StatusLink/StatusLink';
+import { FlattenService } from '../Inventory.types';
 
-import {
-  getBadgeColorForServiceStatus,
-  getBadgeIconForServiceStatus,
-  getBadgeTextForServiceStatus,
-  getAgentsMonitoringStatus,
-  getNodeLink,
-} from './Services.utils';
+import { getAgentsMonitoringStatus } from './Services.utils';
+import ServicesTable from './Services/ServicesTable';
 import { getStyles } from './Tabs.styles';
 
 export const Services = () => {
@@ -67,129 +51,6 @@ export const Services = () => {
         };
       }),
     [fetchedServices]
-  );
-
-  const getActions = useCallback(
-    (row: Row<FlattenService>): Action[] => [
-      {
-        content: (
-          <HorizontalGroup spacing="sm">
-            <Icon name="trash-alt" />
-            <span className={styles.deleteItemTxtSpan}>{Messages.delete}</span>
-          </HorizontalGroup>
-        ),
-        action: () => {
-          setActionItem(row.original);
-          setModalVisible(true);
-        },
-      },
-      {
-        content: Messages.services.actions.dashboard,
-        action: () => {
-          locationService.push(getDashboardLinkForService(row.original.type, row.original.serviceName));
-        },
-      },
-      {
-        content: Messages.services.actions.qan,
-        action: () => {
-          locationService.push(`/d/pmm-qan/pmm-query-analytics?var-service_name=${row.original.serviceName}`);
-        },
-      },
-    ],
-    [styles.deleteItemTxtSpan]
-  );
-
-  const columns = useMemo(
-    (): Array<ExtendedColumn<FlattenService>> => [
-      {
-        Header: Messages.services.columns.serviceId,
-        id: 'serviceId',
-        accessor: 'serviceId',
-        hidden: true,
-        type: FilterFieldTypes.TEXT,
-      },
-      {
-        Header: Messages.services.columns.status,
-        accessor: 'status',
-        Cell: ({ value }: { value: ServiceStatus }) => (
-          <Badge
-            text={getBadgeTextForServiceStatus(value)}
-            color={getBadgeColorForServiceStatus(value)}
-            icon={getBadgeIconForServiceStatus(value)}
-          />
-        ),
-        tooltipInfo: <StatusInfo />,
-        type: FilterFieldTypes.DROPDOWN,
-        options: [
-          {
-            label: 'Up',
-            value: ServiceStatus.UP,
-          },
-          {
-            label: 'Down',
-            value: ServiceStatus.DOWN,
-          },
-          {
-            label: 'Unknown',
-            value: ServiceStatus.UNKNOWN,
-          },
-          {
-            label: 'N/A',
-            value: ServiceStatus.NA,
-          },
-        ],
-      },
-      {
-        Header: Messages.services.columns.serviceName,
-        accessor: 'serviceName',
-        Cell: ({ value, row }: { row: Row<FlattenService>; value: string }) => (
-          <ServiceIconWithText text={value} dbType={row.original.type} />
-        ),
-        type: FilterFieldTypes.TEXT,
-      },
-      {
-        Header: Messages.services.columns.nodeName,
-        accessor: 'nodeName',
-        Cell: ({ value, row }: { row: Row<FlattenService>; value: string }) => (
-          <Link className={styles.link} href={getNodeLink(row.original)}>
-            {value}
-          </Link>
-        ),
-        type: FilterFieldTypes.TEXT,
-      },
-      {
-        Header: Messages.services.columns.monitoring,
-        accessor: 'agentsStatus',
-        width: '70px',
-        Cell: ({ value, row }) => (
-          <StatusLink type="services" strippedId={stripServiceId(row.original.serviceId)} agentsStatus={value} />
-        ),
-        type: FilterFieldTypes.RADIO_BUTTON,
-        options: [
-          {
-            label: MonitoringStatus.OK,
-            value: MonitoringStatus.OK,
-          },
-          {
-            label: MonitoringStatus.FAILED,
-            value: MonitoringStatus.FAILED,
-          },
-        ],
-      },
-      {
-        Header: Messages.services.columns.address,
-        accessor: 'address',
-        type: FilterFieldTypes.TEXT,
-      },
-      {
-        Header: Messages.services.columns.port,
-        accessor: 'port',
-        width: '100px',
-        type: FilterFieldTypes.TEXT,
-      },
-      getExpandAndActionsCol(getActions),
-    ],
-    [styles, getActions]
   );
 
   const deletionMsg = useMemo(() => {
@@ -257,40 +118,10 @@ export const Services = () => {
     setSelectedRows(rows);
   }, []);
 
-  const renderSelectedSubRow = React.useCallback(
-    (row: Row<FlattenService>) => {
-      const labels = row.original.customLabels || {};
-      const labelKeys = Object.keys(labels);
-      const agents = row.original.agents || [];
-
-      return (
-        <DetailsRow>
-          {!!agents.length && (
-            <DetailsRow.Contents title={Messages.services.details.agents}>
-              <StatusBadge
-                strippedId={stripServiceId(row.original.serviceId)}
-                type={'services'}
-                agents={row.original.agents || []}
-              />
-            </DetailsRow.Contents>
-          )}
-          <DetailsRow.Contents title={Messages.services.details.serviceId}>
-            <span>{row.original.serviceId}</span>
-          </DetailsRow.Contents>
-          {!!labelKeys.length && (
-            <DetailsRow.Contents title={Messages.services.details.labels} fullRow>
-              <TagList
-                colorIndex={9}
-                className={styles.tagList}
-                tags={labelKeys.map((label) => `${label}=${labels![label]}`)}
-              />
-            </DetailsRow.Contents>
-          )}
-        </DetailsRow>
-      );
-    },
-    [styles.tagList]
-  );
+  const handleDelete = useCallback((service: FlattenService) => {
+    setActionItem(service);
+    setModalVisible(true);
+  }, []);
 
   const onModalClose = useCallback(() => {
     setModalVisible(false);
@@ -357,22 +188,11 @@ export const Services = () => {
               )}
             />
           </Modal>
-          <Table
-            columns={columns}
-            data={flattenServices}
-            totalItems={flattenServices.length}
-            rowSelection
-            onRowSelection={handleSelectionChange}
-            showPagination
-            pageSize={25}
-            allRowsSelectionMode="page"
-            emptyMessage={Messages.services.emptyTable}
-            pendingRequest={isLoading}
-            overlayClassName={styles.overlay}
-            renderExpandedRow={renderSelectedSubRow}
-            autoResetSelectedRows={false}
-            getRowId={useCallback((row: FlattenService) => row.serviceId, [])}
-            showFilter
+          <ServicesTable
+            flattenServices={flattenServices}
+            onSelectionChange={handleSelectionChange}
+            onDelete={handleDelete}
+            isLoading={isLoading}
           />
         </FeatureLoader>
       </OldPage.Contents>
